requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.9.3
idna==2.6
requests==2.18.4
urllib3==1.22
mysqlclient==1.3.12
djangorestframework==3.7.7